requirements.txt
fastapi
uvicorn
requests
openai
psycopg2-binary
python-dotenv